chore: rename shared to common in root
