+++ .eslintrc.js
@@ -48,6 +48,7 @@ module.exports = {
       rules: {
         '@typescript-eslint/explicit-module-boundary-types': 0,
         '@typescript-eslint/no-non-null-assertion': 0,
+        '@typescript-eslint/no-explicit-any': 0,
       },
     },
   ],
