+++ src/common/shared.ts
@@ -25,7 +25,7 @@ THE SOFTWARE.
 import type { Step, Steps } from '@elastic/synthetics';
 import { RendererProcessIpc } from 'electron-better-ipc';
 import React from 'react';
-import type { Journey, JourneyType } from '../../shared/types';
+import type { Journey, JourneyType } from '../../common/types';
 
 export const COMMAND_SELECTOR_OPTIONS = [
   {
